修正：阿里云主机清理回收站功能

+++ pkg/compute/options/options.go
@@ -22,8 +22,8 @@ type ComputeOptions struct {
 	DefaultDiskSize int `default:"30720" help:"Default disk size in MB if not specified, default to 30GiB"`
 
 	EnablePendingDelete            bool  `default:"true" help:"Turn on/off pending delete VM and disk, default is on"`
-	PendingDeleteExpireSeconds     int64 `default:"259200" help:"How long a pending delete VM cleaned automatically, default 3 days"`
-	PendingDeleteMaxCleanBatchSize int32 `default:"50" help:"How many pending delete servers can be clean in a batch"`
+	PendingDeleteExpireSeconds     int `default:"259200" help:"How long a pending delete VM cleaned automatically, default 3 days"`
+	PendingDeleteMaxCleanBatchSize int `default:"50" help:"How many pending delete servers can be clean in a batch"`
 
 	ImageCacheStoragePolicy string `default:"least_used" choices:"best_fit|least_used" help:"Policy to choose storage for image cache, best_fit or least_used"`
 	MetricsRetentionDays    int32  `default:"30" help:"Retention days for monitoring metrics in influxdb"`
